% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_f.R
\name{aest.f}
\alias{aest.f}
\title{Assign point color and shape aesthetics.}
\usage{
aest.f(x, n.cat = 2, option = "D")
}
\arguments{
\item{x}{Character vector with labels, or a numerical vector to be 
discretized in 'n.cat' categories.}

\item{n.cat}{Number of categories to split vector 'x'. Numeric.
Ignored if 'x' is a character vector.}

\item{option}{Controls color palette.
One of the possible 'option' arguments for the 'viridis' function.}
}
\value{
A data.frame with labels as rownames and two
columns representing point colors and shape, respectively.
}
\description{
This function is called by moss whenever a plot is produced.
It simply assigns colors and shape to points based on input labels.
}
