% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_preprocess.R
\name{check_and_preprocess_data}
\alias{check_and_preprocess_data}
\title{Validate and preprocess the data}
\usage{
check_and_preprocess_data(dat, dW, nbin = 4, T.cutoff = NULL)
}
\arguments{
\item{dat}{The dat}

\item{dW}{The d w}

\item{nbin}{how many levels should continuous W's be binned into}
}
\description{
Validate and preprocess the data
}
