% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_for_onestep.R
\name{compute_onestep_update_matrix}
\alias{compute_onestep_update_matrix}
\title{Perform one-step TMLE update of survival curve}
\usage{
compute_onestep_update_matrix(D1.t.func.prev, Pn.D1.func.prev, dat, T.uniq,
  W_names, dW)
}
\arguments{
\item{D1.t.func.prev}{n*p matrix of previous influence curve}

\item{Pn.D1.func.prev}{p vector of previous mean influence curve}

\item{dat}{input data.frame}

\item{T.uniq}{grid of unique event times}

\item{W_names}{vector of the names of baseline covariates}

\item{dW}{dynamic intervention}
}
\value{
vector of length n_sample
}
\description{
Perform one-step TMLE update of survival curve
}
