% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{compute_q}
\alias{compute_q}
\title{compute multivariate normal quantile from a correlation matrix}
\usage{
compute_q(corr, B = 1000, alpha = 0.05)
}
\arguments{
\item{corr}{correlation matrix}

\item{B}{number of monte-carlo samples drawn to estimate the quantile}

\item{alpha}{significant level (to compute 1-alpha quantile)}
}
\value{
univariate numeric quantile of the quantile(max_j(abs(x))) where x is
 drawn from normal(0, corr)
}
\description{
compute multivariate normal quantile from a correlation matrix
}
