% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{compute_simultaneous_ci}
\alias{compute_simultaneous_ci}
\title{compute simutaneous confidence band around a survival curve esimator}
\usage{
compute_simultaneous_ci(eic_fit)
}
\arguments{
\item{eic_fit}{a matrix of efficient influence curve from `eic` class
`all_t` method}
}
\value{
a vector of standard error corresponding to each time point on the
survival curve
}
\description{
compute simutaneous confidence band around a survival curve esimator
}
