% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_for_onestep.R
\name{create_step_func}
\alias{create_step_func}
\title{generate step function object using y value and corresponding jump points}
\usage{
create_step_func(y.vec, t.vec)
}
\arguments{
\item{y.vec}{a vector of step function values}

\item{t.vec}{a vector for all jump locations of step function.
NOTE: the first element value of y.vec is the flat part LEFT of the first jump point specified in t.vec}
}
\value{
sfun a function object, that can perform any mapping
}
\description{
generate step function object using y value and corresponding jump points
}
