% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_for_onestep.R
\name{create_Yt_vector}
\alias{create_Yt_vector}
\title{compute \eqn{I\{T.tilde >= t\}}}
\usage{
create_Yt_vector(Time, t.vec)
}
\arguments{
\item{Time}{length n vector of failure time}

\item{t.vec}{t value of interest}
}
\value{
a binary vector, of length = t.vec
}
\description{
loop over t.vec
}
