% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_for_onestep.R
\name{create_Yt_vector_with_censor}
\alias{create_Yt_vector_with_censor}
\title{compute \eqn{I\{T.tilde == t, Delta = 1\}}}
\usage{
create_Yt_vector_with_censor(Time, Delta, t.vec)
}
\arguments{
\item{Time}{length n vector of failure time}

\item{Delta}{length n vector of censoring indicator}

\item{t.vec}{t value of interest}
}
\value{
a binary vector, of length = t.vec
}
\description{
loop over t.vec
}
