% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_SL_fit.R
\name{estimate_censoring_SL}
\alias{estimate_censoring_SL}
\title{wrapper for censoring regression SuperLearner}
\usage{
estimate_censoring_SL(dat, T.uniq, Delta.SL.Lib = c("SL.mean", "SL.glm",
  "SL.gam", "SL.earth"))
}
\arguments{
\item{dat}{data.frame, with col: T, A, W, Delta}

\item{T.uniq}{the unique time points of failure}

\item{Delta.SL.Lib}{library for SuperLearner}
}
\value{
h.hat.t matrix, nrow = n.sample, ncol = length(T.uniq);
each row is the predicted conditional hazard h(T=t | T>=t, A, W) for that subject
}
\description{
wrapper for censoring regression SuperLearner
}
\examples{
# TO DO
}
