% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_SL_fit.R
\name{estimate_hazard_SL}
\alias{estimate_hazard_SL}
\title{wrapper for conditonal hazard regression SuperLearner}
\usage{
estimate_hazard_SL(dat, T.uniq, ht.SL.Lib = c("SL.mean", "SL.glm", "SL.gam",
  "SL.earth"))
}
\arguments{
\item{dat}{data.frame, with col: T, A, C, W}

\item{T.uniq}{vector of unique event times}

\item{ht.SL.Lib}{library for SuperLearner}
}
\value{
h.hat.t matrix, nrow = n.sample, ncol = length(T.uniq);
         each row is the predicted conditional hazard h(T=t | T>=t, A, W) for that subject
}
\description{
use T, A, C, W data format as input
}
\examples{
# TO DO
}
