% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_SL_fit.R
\name{estimateHazards}
\alias{estimateHazards}
\title{Conditional Hazard estimation in survtmle package
From survtmle (David Benkeser, 2016)}
\usage{
estimateHazards(dataList, J, verbose, strata = NULL, adjustVars,
  SLlibrary.event, glmFormula.event, superLearnerSummary, bounds)
}
\arguments{
\item{dataList}{output of makeDataList}

\item{J}{same as above}

\item{verbose}{print messages as SL runs?}

\item{strata}{would leave as NULL, dont remember exactly how it is coded}

\item{adjustVars}{data.frame of predictors for SL, dont know why I have it coded this way...
just put in e.g. dat[,c("W1","W2")]. by default it will add the Z column
and t column in dataList[[1]] when it fits the SL}

\item{SLlibrary.event}{Super Learner library}

\item{glmFormula.event}{a glm formula if you don't want to use SL}

\item{superLearnerSummary}{doesn't do anything, just leave blank}

\item{bounds}{NULL}
}
\value{
added column for the initial data.frame
QjHaz <- result, for each subject, h_t from time 0 to T-1; organized in T_i rows
QjPseudoHaz <-
returns list of length = length(dataList) with added columns name QjPseudoHaz and
QjHaz for all j in J. Ignore the PseudoHaz variables -- they're used when there are
multiple event types; if there's only one type of event they'll == Haz values anyway
}
\description{
Conditional Hazard estimation in survtmle package
From survtmle (David Benkeser, 2016)
}
\examples{
# TO DO
}
