% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_curve.R
\docType{class}
\name{evaluate_metric}
\alias{evaluate_metric}
\title{evaluate one survival_curve against another as truth}
\format{\code{\link{R6Class}} object.}
\usage{
evaluate_metric
}
\value{
Object of \code{\link{R6Class}} with methods
}
\description{
evaluate one survival_curve against another as truth
}
\section{Fields}{

\describe{
\item{\code{survival}}{estimated survival}

\item{\code{survival_truth}}{true survival}
}}

\section{Methods}{

evaluate_mse create data.frame that computes mse for each time point
}

\examples{
\donttest{
  evaluation <- evaluate_metric$new(survival, survival_truth)
  evaluation$evaluate_mse()
}
}
\keyword{data}
