% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet_constrained.R
\name{fit_enet_constrained}
\alias{fit_enet_constrained}
\title{logistic elastic net regression (constrained form)}
\usage{
fit_enet_constrained(Y, X, beta_init, norm_max, offset = NULL,
  type = "l2")
}
\arguments{
\item{Y}{vector of binary outcome}

\item{X}{matrix of predictor}

\item{beta_init}{initial value for slope}

\item{norm_max}{upper constraint on l2/l1 norm of slope}

\item{offset}{a vector of constant offset in the linear term}

\item{type}{fit ridge regression ("l2") or lasso ("l1")}
}
\value{
the fitted slope
}
\description{
logistic elastic net regression (constrained form)
}
