% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_constrained.R
\name{fit_ridge_constrained}
\alias{fit_ridge_constrained}
\title{logistic ridge regression (constrained form)}
\usage{
fit_ridge_constrained(Y, X, beta_init, l2_norm_max, offset = NULL)
}
\arguments{
\item{Y}{vector of binary outcome}

\item{X}{matrix of predictor}

\item{beta_init}{initial value for slope}

\item{l2_norm_max}{upper constraint on l2 norm of slope}

\item{offset}{a vector of constant offset in the linear term}
}
\value{
the fitted slope
}
\description{
logistic ridge regression (constrained form)
}
