% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_fit_survival.R
\name{initial_sl_fit}
\alias{initial_sl_fit}
\title{super learner fit for failure and censoring event}
\usage{
initial_sl_fit(T_tilde, Delta, A, W, t_max, sl_failure = c("SL.glm"),
  sl_censoring = c("SL.glm"), sl_treatment = c("SL.glm"),
  gtol = 0.001)
}
\arguments{
\item{T_tilde}{vector of last follow up time}

\item{Delta}{vector of censoring indicator}

\item{A}{vector of treatment}

\item{W}{data.frame of baseline covariates}

\item{t_max}{the maximum time to estimate the survival probabilities}

\item{sl_failure}{SuperLearner library for failure event hazard}

\item{sl_censoring}{SuperLearner library for censoring event hazard}

\item{sl_treatment}{SuperLearner library for propensity score}

\item{gtol}{treshold for the fitted propensity scores}
}
\description{
using survtmle package
}
