% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw_ee.R
\docType{class}
\name{ipcw}
\alias{ipcw}
\title{run ipcw based on initial fit}
\format{\code{\link{R6Class}} object.}
\usage{
ipcw
}
\value{
Object of \code{\link{R6Class}} with methods
}
\description{
run ipcw based on initial fit
}
\section{Fields}{

\describe{
\item{\code{A}}{vector of treatment}

\item{\code{T_tilde}}{vector of last follow up time}

\item{\code{Delta}}{vector of censoring indicator}

\item{\code{density_failure}}{survival_curve object of predicted counterfactual
survival curve}

\item{\code{density_censor}}{survival_curve object of predicted counterfactual
failure event survival curve}

\item{\code{g1W}}{propensity score}

\item{\code{A_intervene}}{the intervention of interest}

\item{\code{k_grid}}{vector of interested time points}
}}

\section{Methods}{

fit fit one time point
}

\examples{
# ipcw$new(A, T_tilde, Delta, density_failure, density_censor, g1W, A_intervene)
}
\keyword{data}
