% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_for_onestep.R
\name{l2_inner_prod_step}
\alias{l2_inner_prod_step}
\title{compute l2 inner product of two step functions}
\usage{
l2_inner_prod_step(f.step, g.step, T.grid)
}
\arguments{
\item{f.step}{two step-function pdf with shared jump points; can be matrix input: nrow = # of different step-function pdf, ncol = length(T.grid)}

\item{g.step}{two step-function pdf with shared jump points; can be matrix input: nrow = # of different step-function pdf, ncol = length(T.grid)}

\item{T.grid}{shared jump points}
}
\value{
scalar
}
\description{
f and g
}
