% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_SL_fit.R
\name{makeDataList}
\alias{makeDataList}
\title{generate data.frame to be used in survtmle package}
\usage{
makeDataList(dat, J, nZ, Z, t0, bounds = NULL)
}
\arguments{
\item{dat}{a data.frame with columns named:
id = unique subject id
T.tilde = (possibly censored) failure time
Delta.J= (possible censored) failure type -- can set all to 1 if no censoring
Z = binary treatment indicator
W = other columns for baseline variables}

\item{J}{the unique positive values of dat$Delta.J (set J=1 if only one type of failure)}

\item{nZ}{set to length of unique(dat$Z)}

\item{Z}{the treatment values that are of interest (e.g., Z=c(0,1))}

\item{t0}{the time point you eventually want predictions at}

\item{bounds}{just leave NULL}
}
\description{
generate data.frame to be used in survtmle package
}
\examples{
# TO DO
}
