% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metdat.R
\name{metdat}
\alias{metdat}
\title{Extracts (and merges) chunks of characters.}
\usage{
metdat(x, i, sep = "-", collapse = sep)
}
\arguments{
\item{x}{A character vector.}

\item{i}{Index specifying which chunks of characters 
will be extracted (and merged).}

\item{sep}{Chunks separator character. Defaults to "-".}

\item{collapse}{New chunks separator character.
Default to 'sep'.}
}
\value{
A character vector with the extracted (and merged) chunks of
characters.
}
\description{
Extracts (and merges) chunks of characters.
}
\examples{
x <- "this is one chunk of characters & this is another one"
metdat(x, 1, " & ")
metdat(x, 2, " & ")
metdat(x, c(1, 2), " & ")
metdat(x, c(1, 2), " & ", " and ")
}
