% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moss_heatmap.R
\name{moss_heatmap}
\alias{moss_heatmap}
\title{Creates a heatmap from the output of MOSS.}
\usage{
moss_heatmap(B, SVD, right.lab, left.lab, axes.pos = NULL, verbose = TRUE)
}
\arguments{
\item{B}{An bject of class 'matrix' or 'FBM'.}

\item{SVD}{List with the results of a sparse or dense SVD.}

\item{right.lab}{Columns title. Character.}

\item{left.lab}{Rows title. Character.}

\item{axes.pos}{What SVD dimensions should be used to plot the heatmap? 
If NULL, all the SVD dimensions are use. Defaults to NULL.}

\item{verbose}{Should we print messages? Logical.
Defaults to TRUE.}
}
\value{
Returns a 'ComplexHeatmap' plot representing the cross-product
 between left and right Eigenvectors.
}
\description{
Creates a heatmap from the output of MOSS.
}
