% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moss_venn.R
\name{moss_venn}
\alias{moss_venn}
\title{Useful Venn diagrams to study the overlap between samples row names.}
\usage{
moss_venn(L, a, lty = "blank", fill = NULL, element_names = NULL)
}
\arguments{
\item{L}{List of elements which overlap we wish to check (e.g., row names 
by omic blocks).}

\item{a}{Elements of the list we want to focus on (e.g., a subset of omic 
blocks). Numerical.}

\item{lty}{Line width of circles circumferences.}

\item{fill}{Color for each circle. Character vector. Defaults to NULL.}

\item{element_names}{Names of each category. Character vector.
Defaults to NULL}
}
\value{
A data.frame with labels as rownames and two
}
\description{
Useful Venn diagrams to study the overlap between samples row names.
}
