% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.surv_onestep}
\alias{plot.surv_onestep}
\title{Plot the survival curve estimator}
\usage{
\method{plot}{surv_onestep}(onestepfit, col = "green", add = FALSE, ...)
}
\arguments{
\item{onestepfit}{object returned by surv_onestep or surv_onestep_complete}

\item{col}{line color}

\item{add}{whether to add to existing plot}

\item{...}{}
}
\value{
NA
}
\description{
Plot the survival curve estimator
}
