% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.surv_survtmle}
\alias{plot.surv_survtmle}
\title{Plot the survival curve estimator}
\usage{
\method{plot}{surv_survtmle}(obj, add = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{
NA
}
\description{
Plot the survival curve estimator
}
