% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro_na.R
\name{prepro_na}
\alias{prepro_na}
\title{Number of missing values within a matrix.}
\usage{
prepro_na(X)
}
\arguments{
\item{X}{An object of class 'matrix', 'FBM', or 'array'.}
}
\value{
Returns total number of missing values in X.
}
\description{
This function is called by moss to count the number of missing values within the (extended) matrices.
}
\details{
Ment for objects of class 'matrix', 'FBM', or 'array'.
}
