% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro_sub.R
\name{prepro_sub}
\alias{prepro_sub}
\title{Scale and normalize columns of a matrix.}
\usage{
prepro_sub(X, scale.arg, norm.arg)
}
\arguments{
\item{X}{An object of class 'matrix', 'FBM', or 'array'.}

\item{scale.arg}{Should we scale columns? Logical.}

\item{norm.arg}{Should we normalize columns? Logical.}
}
\value{
A matrix with scaled and/or normalized columns.
}
\description{
This function is called by moss to scale and normalize (extended) matrices.
}
\details{
Ment for objects of class 'matrix', 'FBM', or 'array'.
}
