% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_methods.R
\name{summarize_weights}
\alias{summarize_weights}
\title{summarize propensity score weights from surv_onestep object}
\usage{
summarize_weights(onestep_curve)
}
\arguments{
\item{onestep_curve}{}
}
\description{
summarize propensity score weights from surv_onestep object
}
\examples{
# NA
}
