% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_curve.R
\docType{class}
\name{survival_curve}
\alias{survival_curve}
\title{general object to hold hazard, survival, or pdf}
\format{\code{\link{R6Class}} object.}
\usage{
survival_curve
}
\value{
Object of \code{\link{R6Class}} with methods
}
\description{
general object to hold hazard, survival, or pdf
}
\section{Fields}{

\describe{
\item{\code{t}}{vector of time points}

\item{\code{hazard}}{matrix or vector of hazard}

\item{\code{survival}}{matrix or vector of survival}

\item{\code{pdf}}{matrix or vector of pdf}
}}

\section{Methods}{

display plot the conditonal survival/hazard/pdf
}

\examples{
\donttest{
   survival_curve$new(t, hazard)
   survival_curve$new(t, survival)
   survival_curve$new(t, pdf)
}
}
\keyword{data}
