% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_survtmle.R
\name{survtmle_single_t}
\alias{survtmle_single_t}
\title{Iterative TMLE for survival curve at specific time point}
\usage{
survtmle_single_t(dat, tk, dW = rep(1, nrow(dat)), T.cutoff = NULL,
  SL.ftime = c("SL.glm", "SL.mean", "SL.step", "SL.earth"),
  SL.ctime = c("SL.glm", "SL.mean"), SL.trt = c("SL.glm", "SL.mean",
  "SL.step", "SL.earth"))
}
\arguments{
\item{dat}{data.frame with columns T, A, C, W. All columns with character "W" will be treated as baseline covariates.}

\item{tk}{time point to compute survival probability}

\item{SL.trt}{}
}
\description{
Iterative TMLE for survival curve at specific time point
}
