% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\docType{class}
\name{eic}
\alias{eic}
\title{compute eic under the likelihood fit for each observation}
\format{\code{\link{R6Class}} object.}
\usage{
eic
}
\value{
Object of \code{\link{R6Class}} with methods
}
\description{
compute eic under the likelihood fit for each observation
}
\section{Fields}{

\describe{
\item{\code{A}}{vector of treatment}

\item{\code{T_tilde}}{vector of last follow up time}

\item{\code{Delta}}{vector of censoring indicator}

\item{\code{density_failure}}{survival_curve object of predicted counterfactual
survival curve}

\item{\code{density_censor}}{survival_curve object of predicted counterfactual
failure event survival curve}

\item{\code{g1W}}{propensity score}

\item{\code{psi}}{a vector of target parameter estimate}

\item{\code{A_intervene}}{the intervention of interest}

\item{\code{k_grid}}{vector of interested time points}
}}

\section{Methods}{

one_t compute a vector of EIC
all_t compute a matrix of EIC for all time points
clever_covariate compute the clever covariate for one time point
}

\keyword{data}
