% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_fit_survival.R
\name{initial_sl_fit}
\alias{initial_sl_fit}
\title{super learner fit for failure and censoring event}
\usage{
initial_sl_fit(ftime, ftype, trt, adjustVars, t_0, trtOfInterest = 0:1,
  SL.trt = c("SL.glm"), gtol = 0.001, SL.ctime = c("SL.glm"),
  SL.ftime = c("SL.glm"))
}
\arguments{
\item{ftime}{vector of last follow up time}

\item{ftype}{vector of censoring indicator}

\item{trt}{vector of treatment}

\item{adjustVars}{data.frame of baseline covariates}

\item{t_0}{the maximum time to estimate the survival probabilities}

\item{trtOfInterest}{the intervention of interest}

\item{SL.trt}{SuperLearner library for propensity score}

\item{gtol}{treshold for the fitted propensity scores}

\item{SL.ctime}{SuperLearner library for censoring event hazard}

\item{SL.ftime}{SuperLearner library for failure event hazard}
}
\description{
using survtmle package
}
