% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moss_hazard.R
\docType{class}
\name{MOSS_hazard}
\alias{MOSS_hazard}
\title{onestep TMLE}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MOSS_hazard
}
\description{
onestep TMLE
}
\keyword{datasets}
