% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moss_ate.R
\docType{class}
\name{MOSS_hazard_ate}
\alias{MOSS_hazard_ate}
\title{onestep TMLE of average treatment effect on survival probabilities}
\format{\code{\link{R6Class}} object.}
\usage{
MOSS_hazard_ate
}
\value{
Object of \code{\link{R6Class}} with methods
}
\description{
updating the hazard using constrained step size update
}
\section{Fields}{

\describe{
\item{\code{A}}{vector of treatment}

\item{\code{T_tilde}}{vector of last follow up time}

\item{\code{Delta}}{vector of censoring indicator}

\item{\code{density_failure}}{survival_curve object of predicted counterfactual
survival curve}

\item{\code{density_censor}}{survival_curve object of predicted counterfactual
failure event survival curve}

\item{\code{density_failure_0}}{survival_curve object of predicted counterfactual
survival curve 0}

\item{\code{density_censor_0}}{survival_curve object of predicted counterfactual
failure event survival curve 0}

\item{\code{g1W}}{propensity score}

\item{\code{k_grid}}{vector of interested time points}
}}

\section{Methods}{

iterate_onestep update the initial estimator
}

\examples{
\donttest{
  MOSS_hazard_ate$new(
    A,
    T_tilde,
    Delta,
    density_failure,
    density_censor,
    density_failure_0,
    density_censor_0,
    g1W,
    A_intervene = 1,
    k_grid = 1:max(T_tilde)
  )
}
}
\keyword{data}
