% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.surv_survtmle}
\alias{plot.surv_survtmle}
\title{Plot the survival curve estimator}
\usage{
\method{plot}{surv_survtmle}(obj, add = FALSE, ...)
}
\arguments{
\item{add}{whether to add to existing plot}

\item{...}{}

\item{onestepfit}{object returned by surv_onestep or surv_onestep_complete}

\item{col}{line color}

\item{...}{}
}
\value{
NA

NA
}
\description{
Plot the survival curve estimator
}
\examples{
Plot the survival curve estimator

}
