% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro_na.R
\name{prepro_na}
\alias{prepro_na}
\title{Missing values imputation by the mean of each column.}
\usage{
prepro_na(X)
}
\arguments{
\item{X}{An object of class 'matrix', 'FBM', or 'array'.}
}
\value{
Returns input with imputed missing values.
}
\description{
This function is called by moss to count the impute missing values
by the mean of each column within omic blocks.
If any column has more than 20% missing values an error is displayed.
}
\details{
Meant for objects of class 'matrix', 'FBM', or 'array'.
}
