% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw_ee.R
\docType{class}
\name{repeat_t_grid}
\alias{repeat_t_grid}
\title{repeat ipcw/ee for all time points}
\format{\code{\link{R6Class}} object.}
\usage{
repeat_t_grid
}
\value{
Object of \code{\link{R6Class}} with methods
}
\description{
repeat ipcw/ee for all time points
}
\section{Fields}{

\describe{
\item{\code{method}}{either ipcw or ee class}

\item{\code{A}}{vector of treatment}

\item{\code{T_tilde}}{vector of last follow up time}

\item{\code{Delta}}{vector of censoring indicator}

\item{\code{density_failure}}{survival_curve object of predicted counterfactual
survival curve}

\item{\code{density_censor}}{survival_curve object of predicted counterfactual
failure event survival curve}

\item{\code{g1W}}{propensity score}

\item{\code{A_intervene}}{the intervention of interest}

\item{\code{k_grid}}{vector of interested time points}
}}

\section{Methods}{

fit fit all time points in `k_grid`
}

\examples{
# repeat_t_grid$new(method, A, T_tilde, Delta, density_failure, density_censor, g1W, A_intervene)
}
\keyword{data}
